+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 0.5.0.0 (UNRELEASED)
 
+This release comes with major API refactor. Key changes:
+
+- Support for both Pandoc and MMark parsers
+- Add `Rib.Markup.Markup` type class to polymorphically select the parser
+- Add top-level `Rib` import namespace to ease of use
+- Drop support for JSON cache
+- Replace complex metadata handling using straightforward aeson `Value` parsing
+
 ## 0.4.1.0
 
 - `Rib.Pandoc`: